Create Trigger
Altering database, then setting the user with rollback. Then dropping the database, and recreating the database, and setting the no count option to on, then creating the table, and creating a trigger for the table to print the Date, and inserting the test 1 values into the columns Sou_ID, and Sou_Description.